Seperated Address into new class
